+++ addon/mixins/select-box/option/base.js
@@ -1,7 +1,6 @@
 import Mixin from '@ember/object/mixin';
 import Nameable from  '../../general/nameable';
 import Registerable from  '../../general/registerable';
-import { isBlank } from '@ember/utils';
 import { resolve } from 'rsvp';
 import { bind } from '@ember/runloop';
 
@@ -19,7 +18,7 @@ export default Mixin.create(...mixins, {
   },
 
   _update(value) {
-    const val = resolve(value);
+    const val = this._resolveValue(value);
     const id = this.incrementProperty('valueID');
 
     const success = bind(this, '_resolvedValue', id, false);
@@ -30,18 +29,28 @@ export default Mixin.create(...mixins, {
     val.then(success, failure);
   },
 
+  _resolveValue(value) {
+    this.set('isPending', true);
+    this.set('isRejected', false);
+    this.set('isFulfilled', false);
+    this.set('isSettled', false);
+
+    return resolve(value);
+  },
+
   _resolvedValue(id, failed, value) {
     if (id < this.get('valueID') || this.get('isDestroyed')) {
       return;
     }
 
-    let label = this.get('label');
-
-    if (isBlank(label)) {
-      label = value;
+    if (failed) {
+      this.set('isRejected', true);
+    } else {
+      this.set('isFulfilled', true);
     }
 
-    this.set('internalLabel', label);
+    this.set('isPending', false);
+    this.set('isSettled', true);
     this.set('internalValue', value);
   }
 });